requirements.txt
Django>=4.2
psycopg2-binary
django-environ
requests